add initial binary generator logic for messages

+++ trains/train.go
@@ -5,6 +5,7 @@ import (
 	"github.com/google/uuid"
 )
 
+//go:generate go run ../cmd/generator/main.go -type=Train -output=../message/train_gen.go
 type Train struct {
 	ID           uuid.UUID
 	IsReversing  bool
@@ -23,8 +24,9 @@ const (
 	CarTypePassenger
 )
 
+//go:generate go run ../cmd/generator/main.go -type=TrainCar -output=../message/train_car_gen.go
 type TrainCar struct {
-	X, Y      int
+	X, Y      uint16
 	Direction types.Dir
 	Type      CarType
 }
